Fixed CLion warning

+++ tests/test_thread_pool.cpp
@@ -60,13 +60,13 @@ TestImpl(test_threadpool)
         //concurrency::parallel_for(size_t(0), numbers.size(), [&](int index) {
         //    sum += ptr[index];
         //});
-        printf("ParallelFor  elapsed: %.3fs  result: %ld\n", timer.elapsed(), (int64_t)sum);
+        printf("ParallelFor  elapsed: %.3fs  result: %lld\n", timer.elapsed(), (long long)sum);
 
         timer.start();
         int64_t sum2 = 0;
         for (int i = 0; i < len; ++i)
             sum2 += ptr[i];
-        printf("Singlethread elapsed: %.3fs  result: %ld\n", timer.elapsed(), sum2);
+        printf("Singlethread elapsed: %.3fs  result: %lld\n", timer.elapsed(), (long long)sum2);
     }
 
 } Impl;
